Fix api response parameter type

+++ core/domain/src/main/java/com/igorvd/bitcoincharts/core/domain/formatter/NumberFormatter.kt
@@ -1,6 +1,6 @@
 package com.igorvd.bitcoincharts.core.domain.formatter
 
 interface NumberFormatter {
-    fun format(value: Int): String
+    fun format(value: Long): String
     fun formatCompact(value: Double): String
 }
